Break out visualization system
